Bonus ok with double array

+++ get_next_line_bonus.h
@@ -6,7 +6,7 @@
 /*   By: dtanigaw <dtanigaw@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/09 04:07:38 by dtanigaw          #+#    #+#             */
-/*   Updated: 2021/04/09 05:17:55 by dtanigaw         ###   ########.fr       */
+/*   Updated: 2021/04/09 05:38:35 by dtanigaw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@
 
 char	*ft_alloc(size_t n);
 void	ft_join(char **res, char **s1, char *s2);
-void	ft_memcpy(char *dest, char *src, int n);
+void	ft_memcpy(char *dest, char *src, size_t n);
 int		ft_strchr(char *s, char c);
 size_t	ft_strlen(char *s);
 char	*ft_strsdup(char *src, size_t size);
